requirements.txt
appnope==0.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-Cors==3.0.6
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
mysqlclient==1.3.13
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.6.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.8.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
